android/ios: split metadata provider sources
