+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/providers/TimestampProvider.kt
@@ -1,0 +1,16 @@
+// capture-sdk - bitdrift's client SDK
+// Copyright Bitdrift, Inc. All rights reserved.
+//
+// Use of this source code is governed by a source available license that can be found in the
+// LICENSE file or at:
+// https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
+
+package io.bitdrift.capture.providers
+
+import io.bitdrift.capture.ITimestampProvider
+
+internal class TimestampProvider(
+    private val dateProvider: DateProvider,
+) : ITimestampProvider {
+    override fun timestamp(): Long = dateProvider.invoke().time
+}
