+++ test/platform/swift/unit_integration/core/TimestampProviderControllerTests.swift
@@ -1,0 +1,18 @@
+// capture-sdk - bitdrift's client SDK
+// Copyright Bitdrift, Inc. All rights reserved.
+//
+// Use of this source code is governed by a source available license that can be found in the
+// LICENSE file or at:
+// https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
+
+@testable import Capture
+import CaptureMocks
+import XCTest
+
+final class TimestampProviderControllerTests: XCTestCase {
+    func testUsesCustomDateProvider() {
+        let timestampProvider = TimestampProviderController(dateProvider: MockDateProvider())
+
+        XCTAssertNotEqual(timestampProvider.timestamp(), 0)
+    }
+}
